archive/requirements.txt
biopython
numpy>=1.13.3
scipy
lxml
pandas
Cython
ete3>=3.0.0b35
pyparsing>=2.1.10
pysam>=0.11.2.2
six>=1.10.0
requests>=2.13.0
dendropy>=4.3.0
tqdm>=4.19.1
pyham
pyyaml
multiprocessing_logging
